+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The following breaking changes only affect you if you're using named arguments:
 
 - `BigInteger::modInverse()` now uses `$modulus` as the parameter name
 - `BigInteger::modPow()` now uses `$exponent` and `$modulus` as parameter names
+- `BigDecimal::withPointMovedLeft()` now uses `$places` as the parameter name
+- `BigDecimal::withPointMovedRight()` now uses `$places` as the parameter name
 
 ## [0.14.8](https://github.com/brick/math/releases/tag/0.14.8) - 2026-02-10